Add files via upload
This one is different from others. Here's the tricky part with this one, If you try to create tables inside the primary one you might have a tricky time adjusting the size of the columns, so be sure what you tinker with. But you can do what you want if you know what you are doing.